backend/requirements.txt
# AI Codebase Onboarding - Backend

## Requirements

### Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

### Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

### AI & ML
openai==1.10.0  # Optional: only for OpenAI embeddings
groq==0.4.1  # For LLM inference (free)
sentence-transformers==2.2.2  # For free local embeddings
langchain==0.1.20
langchain-openai==0.0.8
faiss-cpu==1.9.0.post1
tiktoken==0.5.2

### Code Analysis
GitPython==3.1.40
tree-sitter==0.23.0
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.5

### Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
numpy==1.26.3
redis==5.0.1

### Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.0
flake8==7.0.0
mypy==1.8.0
